separated offload config from zero config
